Split `store/` into different files to avoid circular imports
